+++ src/Akeneo/Category/back/Infrastructure/Converter/InternalApi/InternalApiToStd.php
@@ -4,12 +4,9 @@ declare(strict_types=1);
 
 namespace Akeneo\Category\Infrastructure\Converter\InternalApi;
 
-use Akeneo\Category\Application\Converter\AttributeRequirementChecker;
+use Akeneo\Category\Application\Converter\Checker\InternalApiRequirementChecker;
 use Akeneo\Category\Application\Converter\ConverterInterface;
-use Akeneo\Category\Application\Converter\FieldsRequirementChecker;
 use Akeneo\Category\Infrastructure\Exception\ArrayConversionException;
-use Akeneo\Category\Infrastructure\Exception\StructureArrayConversionException;
-use Webmozart\Assert\Assert;
 
 /**
  * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
@@ -31,25 +28,21 @@ use Webmozart\Assert\Assert;
 class InternalApiToStd implements ConverterInterface
 {
     public function __construct(
-        private FieldsRequirementChecker $fieldsChecker,
-        private AttributeRequirementChecker $attributeChecker
+        private InternalApiRequirementChecker $checker
     ) {
     }
 
     /**
-     * @param array{
-     *     properties: PropertyApi,
-     *     attributes: array<string, AttributeCodeApi|AttributeValueApi>
-     * } $data
+     * @param InternalApi $data
      *
-     * @retrun array|StandardInternalApi
+     * @retrun StandardInternalApi
+     *
+     * @throws ArrayConversionException
      */
     public function convert(array $data): array
     {
         // Validate the internal Api data and structure
-        $this->checkArrayStructure($data);
-        $this->checkProperties($data['properties']);
-        $this->checkAttributes($data['attributes']);
+        $this->checker->check($data);
 
         // Normalize
         $convertedData = [];
@@ -58,56 +51,5 @@ class InternalApiToStd implements ConverterInterface
         $convertedData['values'] = $data['attributes'];
 
         return $convertedData;
-    }
-
-    /**
-     * @param array{
-     *     properties: PropertyApi,
-     *     attributes: array<string, AttributeCodeApi|AttributeValueApi>
-     * } $data
-     *
-     * @throws ArrayConversionException
-     */
-    private function checkArrayStructure(array $data): void
-    {
-        $expectedKeys = ['properties', 'attributes'];
-        try {
-            Assert::keyExists($data, 'properties');
-            Assert::keyExists($data, 'attributes');
-        } catch (\InvalidArgumentException $exception) {
-            throw new StructureArrayConversionException(
-                vsprintf('Fields ["%s", "%s"] is expected', $expectedKeys)
-            );
-        }
-    }
-
-    /**
-     * @param array{code: string, labels: array<string, string>} $properties
-     *
-     * @throws ArrayConversionException
-     */
-    private function checkProperties(array $properties): void
-    {
-        $this->fieldsChecker->checkFieldsExist($properties, ['code', 'labels']);
-        $this->fieldsChecker->checkFieldsNotEmpty($properties, ['code']);
-    }
-
-    /**
-     * @param array<string, AttributeCodeApi|AttributeValueApi> $attributes
-     *
-     * @throws ArrayConversionException
-     */
-    private function checkAttributes(array $attributes): void
-    {
-        $this->attributeChecker->checkKeyExist($attributes, 'attribute_codes');
-
-        /** @var $attributeValues array<string, AttributeValueApi}> */
-        $attributeValues = array_filter($attributes, function ($attributeKey) {
-            return $attributeKey !== 'attribute_codes';
-        }, ARRAY_FILTER_USE_KEY);
-
-        $this->attributeChecker->checkAttributeValueKeysExist($attributeValues, $attributes['attribute_codes']);
-
-        $this->attributeChecker->checkAttributeValueArrayStructure($attributeValues);
     }
 }