+++ src/Akeneo/Platform/Job/front/process-tracker/src/feature/components/JobExecutionList/index.ts
@@ -2,3 +2,4 @@ export * from './JobExecutionSearchBar';
 export * from './JobExecutionTable';
 export * from './StatusFilter';
 export * from './TypeFilter';
+export * from './AutomationFilter';